Complex type examples split into separate main files
